[FLINK-20625][pubsub] Align new source with Kafka package names
