+++ flink-connector-gcp-pubsub/src/main/java/org/apache/flink/connector/gcp/pubsub/source/enumerator/PubSubSourceEnumerator.java
@@ -7,20 +7,22 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package org.apache.flink.streaming.connectors.gcp.pubsub.source.enumerator;
+package org.apache.flink.connector.gcp.pubsub.source.enumerator;
 
 import org.apache.flink.api.connector.source.SplitEnumerator;
 import org.apache.flink.api.connector.source.SplitEnumeratorContext;
-import org.apache.flink.streaming.connectors.gcp.pubsub.source.split.PubSubSplit;
+import org.apache.flink.connector.gcp.pubsub.source.PubSubSource;
+import org.apache.flink.connector.gcp.pubsub.source.split.PubSubSplit;
 
 import javax.annotation.Nullable;
 
@@ -28,10 +30,8 @@ import java.io.IOException;
 import java.util.List;
 
 /**
- * The enumerator for the {@link
- * org.apache.flink.streaming.connectors.gcp.pubsub.source.PubSubSource}. It does not do any work
- * discovery as envisioned by FLIP-27 because GCP Pub/Sub hides partitions and other implementation
- * details.
+ * The enumerator for the {@link PubSubSource}. It does not do any work discovery as envisioned by
+ * FLIP-27 because GCP Pub/Sub hides partitions and other implementation details.
  */
 public class PubSubSourceEnumerator implements SplitEnumerator<PubSubSplit, PubSubEnumeratorState> {
     private final SplitEnumeratorContext<PubSubSplit> context;
