+++ flink-connector-gcp-pubsub/src/main/java/org/apache/flink/connector/gcp/pubsub/source/reader/PubSubRecordEmitter.java
@@ -7,21 +7,22 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package org.apache.flink.streaming.connectors.gcp.pubsub.source.reader;
+package org.apache.flink.connector.gcp.pubsub.source.reader;
 
 import org.apache.flink.api.connector.source.SourceOutput;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.connector.base.source.reader.RecordEmitter;
-import org.apache.flink.streaming.connectors.gcp.pubsub.source.split.PubSubSplitState;
+import org.apache.flink.connector.gcp.pubsub.source.split.PubSubSplitState;
 
 /**
  * A custom {@link RecordEmitter} to emit a record which includes the data of the received GCP
