+++ flink-connector-gcp-pubsub/src/main/java/org/apache/flink/connector/gcp/pubsub/source/split/PubSubSplitState.java
@@ -7,16 +7,17 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package org.apache.flink.streaming.connectors.gcp.pubsub.source.split;
+package org.apache.flink.connector.gcp.pubsub.source.split;
 
 /** A stub for the always empty state of a {@link PubSubSplit}. */
 public class PubSubSplitState extends PubSubSplit {}
